requirements.txt
discord.py==2.3.2
watchdog==4.0.0
